- Lauffähig auf 64 Bit Systemen gemacht - Beim Aunlernen von Tasten sollte es keine Abstürze mehr geben
git-svn-id: svn://svn.berlios.de/ffnetdev/trunk@35 1f4bef6d-8e0a-0410-8695-e467da8aaccf

+++ rfb.h
@@ -33,9 +33,10 @@
 #define RFB_H__
 
 // Define the CARD* types
-typedef unsigned long  CARD32;
-typedef unsigned short CARD16;
+typedef unsigned int  CARD32;
+typedef unsigned short int CARD16;
 typedef unsigned char  CARD8;
+    
 
 // Define the port number offsets
 #define FLASH_PORT_OFFSET 5400
